Refactor file parsing into own module to resolve circular deps
