--- galaxy_importer/loaders/extensions_file.py
@@ -1,48 +1,0 @@
-# (c) 2012-2019, Ansible by Red Hat
-#
-# This file is part of Ansible Galaxy
-#
-# Ansible Galaxy is free software: you can redistribute it and/or modify
-# it under the terms of the Apache License as published by
-# the Apache Software Foundation, either version 2 of the License, or
-# (at your option) any later version.
-#
-# Ansible Galaxy is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# Apache License for more details.
-#
-# You should have received a copy of the Apache License
-# along with Galaxy.  If not, see <http://www.apache.org/licenses/>.
-
-import logging
-import os
-import yaml
-
-from galaxy_importer import exceptions as exc
-
-default_logger = logging.getLogger(__name__)
-
-
-class ExtensionsFileLoader:
-    """Load meta/extensions.yml."""
-
-    def __init__(self, collection_path):
-        self.path = os.path.join(collection_path, "meta/extensions.yml")
-        self.data = None
-        self._load()
-
-    def _load(self):
-        if not os.path.exists(self.path):
-            return
-        with open(self.path) as fp:
-            try:
-                self.data = yaml.safe_load(fp)
-            except Exception:
-                raise exc.RuntimeFileError("Error during parsing of extensions.yml")
-
-    def get_extension_dirs(self):
-        if not self.data:
-            return None
-        # TODO: consider check that ext_dir actually exists and fail import if not
-        return [ext["args"]["ext_dir"] for ext in self.data["extensions"]]
